Fix: Implement CRUD operations and error handling for Tag feature
This commit addresses several issues in the Tag feature:

1.  **Corrected ID Type:** Changed the `id` field in `TagDto` from `String` to `Long` to match the `Tag` entity.
2.  **Added DTO Validation:** Added `@NotBlank` validation to `name` and `colorCode` fields in `TagDto`.
3.  **Implemented Service Logic:**
    *   Implemented `getById(Long id)` in `TagServiceImpl`, including `EntityNotFoundException` if the tag is not found.
    *   Implemented `update(TagDto tagDto)` in `TagServiceImpl`, allowing updates to name, description, and colorCode, and throwing `EntityNotFoundException` if the tag to update is not found.
    *   Implemented `delete(Long id)` in `TagServiceImpl`, throwing `EntityNotFoundException` if the tag to delete is not found.
4.  **Updated Service Interface:** Modified `TagService` method signatures for `getById` and `delete` to use `Long` for ID parameters.
5.  **Added Global Error Handling:** Created `RestExceptionHandler` using `@ControllerAdvice` to handle `jakarta.persistence.EntityNotFoundException` and return an HTTP 404 response. This centralizes error handling for "not found" resources.
6.  **Added Unit Tests:** Created `TagServiceImplTest` with JUnit 5 and Mockito to cover the implemented service methods (`getById`, `update`, `delete`), including success and error scenarios (e.g., tag not found).

+++ src/main/java/org/example/komflow/features/contact/service/TagService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface TagService {
     TagDto create (TagDto tag);
     List<TagDto> getAll();
-    TagDto getById(int id);
+    TagDto getById(Long id); // Changed from int to Long
     TagDto update(TagDto tag);
-    void delete(int id);
+    void delete(Long id); // Changed from int to Long
 }
